moved plugins to agents dir
