+++ helios_sim/src/agents/sensors/magnetometer.rs
