+++ helios_sim/src/agents/sensors/plugin_set.rs
