+++ helios_sim/src/agents/sensors/raycasting.rs
