+++ helios_sim/src/agents/vehicles/ackermann/adapter.rs
