+++ helios_sim/src/agents/vehicles/ackermann/components.rs
