+++ helios_sim/src/agents/vehicles/plugin_set.rs
