+++ helios_sim/src/plugins/mod.rs
@@ -8,6 +8,4 @@
 //! `docs/archive/helios_sim/` pending a decision to revive (control metrics and
 //! data logging migrate to the `helios_test` crate) or delete them.
 
-pub mod sensors;
-pub mod vehicles;
 pub mod world;
